backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
# Solución al error de 72 bytes: fijamos bcrypt a 3.2.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic[email]==2.5.3
python-dotenv
scalar-fastapi